src/lib/license-authority/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mhamp1-licenseauthority"
version = "2.1.0"  # LICENSEAUTH UPGRADE: Instant activation + hardware binding + grace period
description = "Secure license generation and validation system with instant activation, device binding, and grace period"
authors = [
  { name = "Matt", email = "mhamp1trading@yahoo.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "cryptography>=42.0.4",
  "keyring>=24.0.0",
  "fastapi>=0.109.1",
  "uvicorn>=0.24.0",
  "pydantic>=2.5.0",
  "sqlalchemy>=2.0.0",
  "pyjwt>=2.8.0",
  "python-jose[cryptography]>=3.4.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.18",
  "redis>=5.0.0",
  "slowapi>=0.1.9",
  "email-validator>=2.0.0"
]

[tool.setuptools.packages.find]
where = ["."]
[project.optional-dependencies]
test = ["pytest"]